+++ NEWS.md
@@ -6,6 +6,10 @@
 - Added 'nickname' mnemonics for each viral sOTU
 - Added SRA-STAT taxonomy analysis
 
+## Bug Fixes
+
+- Fixed `qlen` bug misreporting from palmscan with long sequences
+
 # palmid 0.0.3
 
 - Initial release. `palmid` is a containerized analysis suite and R-package for the classification of **viral RNA-dependent RNA Polymerases (RdRP)** based on the **palmprint** sub-domain and the RNA viral palmprint database [`palmdb`](https://github.com/rcedgar/palmdb).